chore(refactor): refactor image verification packages
