+++ pkg/imageverification/imagedataloader/context.go
