+++ pkg/imageverification/imagedataloader/loader.go
