+++ pkg/imageverification/imagedataloader/loader_test.go
