+++ pkg/imageverification/imagedataloader/options.go
