+++ pkg/imageverification/imagedataloader/utils.go
