+++ pkg/imageverification/imageverifiers/cosign/verifier.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/go-logr/logr"
 	policiesv1alpha1 "github.com/kyverno/kyverno/api/policies.kyverno.io/v1alpha1"
-	"github.com/kyverno/kyverno/pkg/imagedataloader"
+	"github.com/kyverno/kyverno/pkg/imageverification/imagedataloader"
 	"github.com/kyverno/kyverno/pkg/logging"
 	"github.com/pkg/errors"
 	"github.com/sigstore/cosign/v2/pkg/cosign"
