+++ pkg/imageverification/imageverifiers/notary/log.go
